chore: Move scripts outside of src
