+++ scripts/followers.js
@@ -1,18 +1,16 @@
 const { inspect } = require('node:util');
 const { SuperfaceClient } = require('@superfaceai/one-sdk');
-const { withAccessToken } = require('../tokens-utils');
+const { withAccessToken } = require('../utils/tokens-utils');
 
-const publishPost = async (message) => {
+const printFollowers = async (profileId) => {
   const sdk = new SuperfaceClient();
 
   try {
     const provider = await sdk.getProvider('twitter');
-    const profile = await sdk.getProfile('social-media/publish-post');
+    const profile = await sdk.getProfile('social-media/followers');
     const result = await withAccessToken((accessToken) =>
-      profile.getUseCase('PublishPost').perform(
-        {
-          text: message,
-        },
+      profile.getUseCase('GetFollowers').perform(
+        { profileId },
         {
           provider,
           parameters: {
@@ -21,12 +19,13 @@ const publishPost = async (message) => {
         }
       )
     );
+
     console.log(inspect(result.unwrap(), false, Infinity, true));
   } catch (err) {
     console.error(inspect(err, false, Infinity, true));
   }
 };
 
-const message = process.argv[2] || 'Hello from Superface.';
+const profileId = process.argv[2] || '1196797704015400960'; // @superfaceai
 
-publishPost(message);
+printFollowers(profileId);
