+++ scripts/publish-post.js
@@ -1,16 +1,18 @@
 const { inspect } = require('node:util');
 const { SuperfaceClient } = require('@superfaceai/one-sdk');
-const { withAccessToken } = require('../tokens-utils');
+const { withAccessToken } = require('../utils/tokens-utils');
 
-const printFollowers = async (profileId) => {
+const publishPost = async (message) => {
   const sdk = new SuperfaceClient();
 
   try {
     const provider = await sdk.getProvider('twitter');
-    const profile = await sdk.getProfile('social-media/followers');
+    const profile = await sdk.getProfile('social-media/publish-post');
     const result = await withAccessToken((accessToken) =>
-      profile.getUseCase('GetFollowers').perform(
-        { profileId },
+      profile.getUseCase('PublishPost').perform(
+        {
+          text: message,
+        },
         {
           provider,
           parameters: {
@@ -19,13 +21,12 @@ const printFollowers = async (profileId) => {
         }
       )
     );
-
     console.log(inspect(result.unwrap(), false, Infinity, true));
   } catch (err) {
     console.error(inspect(err, false, Infinity, true));
   }
 };
 
-const profileId = process.argv[2] || '1196797704015400960'; // @superfaceai
+const message = process.argv[2] || 'Hello from Superface.';
 
-printFollowers(profileId);
+publishPost(message);
